+++ poc/classes/AuxSTClassInstance.py
@@ -19,3 +19,7 @@ class AuxSTClassInstance(AuxSTInstance):
         new_class_instance.mandatory = self.mandatory
         new_class_instance._is_inherited = True
         return new_class_instance
+
+    def get_type_signature(self):
+        type_child = AuxSymbolTable.get_child_by_outline(self, AuxSymbolTable.type)
+        return type_child.id